chore: Move settings to dedicated folder
