+++ mobile/lib/common/settings/app_details_screen.dart
