+++ mobile/lib/common/settings/share_logs_screen.dart
